Remove unused DatabagSourceFactory
